+++ src/main.go
@@ -17,7 +17,7 @@ func main() {
 	app := NewApp()
 
 	err := wails.Run(&options.App{
-		Title:            "Chobocho Commander V0.2.2",
+		Title:            "Chobocho Commander V0.2.3",
 		Width:            1280,
 		Height:           800,
 		MinWidth:         900,